pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.0",
    "fastapi>=0.129.0",
    "langchain>=1.0.8",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.0.1",
    "langchain-nomic>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.3",
    "pypdf>=6.3.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "streamlit>=1.51.0",
    "uvicorn>=0.41.0",
    "werkzeug>=3.1.5",
]
